requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
jinja2
python-multipart
passlib[bcrypt]
itsdangerous
cryptography
